requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0
pydantic>=2.0.0
python-multipart>=0.0.6
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
dateparser>=1.1.0
openai>=1.3.0  # For LLM summarization
asyncio>=3.4.3
aiofiles>=23.0.0